Unify kody accessor codegen
